moved AppUser, User, and Comments Models to the Models folder
