+++ closeup/ContentView.swift
@@ -5,7 +5,7 @@ struct ContentView: View {
     
     var body: some View {
         ZStack {
-            if let appUser = appUser {
+            if appUser != nil {
                 HomeView(appUser: $appUser)
             } else {
                 SignInView(appUser: $appUser)
