+++ closeup/Models/AppUser.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+struct AppUser: Equatable {
+    let uid: String
+    let email: String?
+}
